created modular code for logistic model
