+++ src/data_preparation.py
@@ -2,7 +2,7 @@ import pandas as pd
 import numpy as np 
 import warnings
 from imblearn.under_sampling import NearMiss
-from sklearn.linear_model import LogisticRegression
+
 from sklearn.metrics import accuracy_score, roc_auc_score
 
 warnings.filterwarnings("ignore")
@@ -11,9 +11,7 @@ from sklearn.impute import SimpleImputer
 from sklearn.preprocessing import LabelEncoder
 
 
-def data_extraction(data_path): 
-    data = pd.read_csv(data_path)
-    return data
+
 
 def data_train_val_test_split(data):
     data.dropna(subset=['RainToday', 'RainTomorrow'], inplace=True)
@@ -67,36 +65,7 @@ def under_sample(train_df):
 
     return x, y 
 
-def train_model(x, y): 
-    model = LogisticRegression(solver='liblinear')
 
-    model.fit(x,y)
-    return model 
-
-def predict(model, X): 
-    return model.predict(X)
-
-def metrics_cal(actuals, predicted): 
-    metrics = {}
-    metrics['auc'] = roc_auc_score(actuals, predicted)
-    metrics['accuracy'] = accuracy_score(actuals, predicted)
-
-    return metrics
-
-
-if __name__ == "__main__": 
-    path = '/Users/mohannaik/Desktop/Mohit_Project/ml-ops-nov-27/data/weatherAUS.csv'
-    data = data_extraction(path)
-    train_df, val_df, test_df = data_train_val_test_split(data)
-    print(train_df.shape)
-    train_df, val_df, test_df = imputing_missing_value(train_df, val_df, test_df)
-    print(train_df.shape)
-    x, y = under_sample(train_df)
-    model = train_model(x, y)
-    test_pred = predict(model, test_df.drop(['RainTomorrow'], axis=1))
-    val_pred = predict(model, val_df.drop(['RainTomorrow'], axis=1))
-    print(metrics_cal(test_df['RainToday'],test_pred))
-    print(metrics_cal(val_df['RainToday'],val_pred))
